W174k: repaired TWO real concurrency breakages; Scene/Fcl +2 MATCH; Camp status +2 MATCH.
INCIDENT 1 -- mdlFile.c, 667 COMPILE_ERROR, now fixed and at 447 MATCH (up from 445). Two distinct corruptions, both from concurrent edits:
  (a) A float-suffix sweep regex appended 'f' to identifiers ending in a digit, producing FUN_00358A30F and FUN_00359380F at 8 call sites. **Any regex adding an 'f' suffix MUST exclude identifier characters before the digit** -- FUN_/DAT_ names end in hex digits and will be eaten.
  (b) An in-flight prototype retype was left HALF-APPLIED: new typed 'extern void FUN_00358A30(float, void*, void*, int)' and FUN_00359380(float, void*, int) were added while the old untyped 'extern u64 FUN_xxx()' declarations survived two lines away, and four call sites still passed raw float BIT PATTERNS (0x40800000, 0x42480000) plus a __int128 by value. Completed the sweep: stale declarations deleted, call sites migrated to 4.0f/50.0f with the address taken on the two Vec128 aggregates. This is the 'retyping a shared prototype REQUIRES sweeping every caller in the same pass' rule failing under concurrency -- a second agent's parallel edit re-introduced the old declaration after the first swept it.
  Recovered clean: no lost MATCHes, over-window 31 -> 30, and 00329A60 and 0034D150 both reached MATCH. 0034D150 finished at 588/592 nd0 by pointing the three post-SQC2 vector reloads at retail's direct $sp+0x40/44/48 operands -- completing a journey from 656/592 nd471.

INCIDENT 2 -- itfMesManager.c compiled to 112 COMPILE_ERROR after two agents wrote it concurrently (LF 6173 -> 6137 -> 6173). Restored byte-exact from the last commit, verified back to 78 MATCH / 34 NONMATCHING, LF-only preserved. Both agents have since serialized between themselves.

Scene/Fcl: mt_sceneFunc 003B8FF0 456/464 nd3 -> MATCH; fclCombineEffect 00417410 320/304 nd172 -> 292/304 MATCH and window cleared; fclCombine 003D32E0 764/736 -> 732/736 and 003D4FF0 1024/1008 -> 1008/1008 windows cleared; mt_scene 003B5620 nd22 -> 19.
Camp: _h_camp_status 001332F0 and 00133780 both MATCH, 41 -> 43.
Recorded negative from Camp: 0013C240 and 0013BE50 were checked against the new unsigned-float-to-int class and neither performs float-to-int at all -- hypothesis excluded for those two.

+++ src/Kosaka/Field/k_field.c
@@ -183,8 +183,8 @@ extern void FUN_004350e0(s32 enabled, s32 update);
 extern void FUN_00429e80(u32 object, s32 enabled);
 extern u32 FUN_001c0330(u8 blockId);
 extern u32 FUN_001bffa0(void);
-extern u64 FUN_001bffe0(void);
-extern u64 FUN_001c0010(void);
+extern u32 FUN_001bffe0(void);
+extern u32 FUN_001c0010(void);
 
 #define FIELD_U8(address)  (*(u8*)(address))
 #define FIELD_U16(address) (*(u16*)(address))
@@ -331,8 +331,8 @@ void* func_001b9480(KwlnTask* fldRootTask)
     u32 request;
     u32 resource;
     u32 camera;
-    u64 list;
-    u64 list2;
+    u32 list;
+    u32 list2;
 
     work = (u8*)fldRootTask->workData;
     state = ROOT_U32(work, 0);
@@ -757,7 +757,7 @@ void* func_001b9480(KwlnTask* fldRootTask)
                     FUN_0017f8d0();
                     FUN_0035bfb0();
                     ROOT_U16(work, 0x14) = FUN_0036f4c0();
-                    FUN_001d03f0((s16)ROOT_U16(work, 0x14));
+                    FUN_001d03f0(ROOT_U16(work, 0x14));
                     FUN_001d70a0();
                     if (ROOT_U16(work, 0x10) < 200)
                     {
